worked on deleting individual tasks and enabling and disabling buttons

+++ script.js
@@ -1,29 +1,24 @@
 const addBtn = document.querySelector('.btn-send');
 const deleteBtn = document.querySelector('.btn-delete');
 const resetBtn = document.getElementById('reset');
+const noBtn = document.getElementById('no-btn');
 
 const inputArea = document.querySelector('.chore-box')
 const error = document.querySelector('.error')
 const taskArea = document.getElementById('task-container');
 const createdTask = document.getElementById('task');
+const modal = document.getElementById('modal');
 
 let tasksArray = [];
 let items = '';
 deleteBtn.disabled = true;
+addBtn.disabled = true;
 
 addBtn.addEventListener('click', () => { addTask(createdTask);});
-deleteBtn.addEventListener('dblclick', () => { removeAllTasks();});
+deleteBtn.addEventListener('click', () => { removeAllTasks();});
 resetBtn.addEventListener('click', () => { reloadList();});
 
-// Added 'Enter' functionality
-
-inputArea.addEventListener('keypress', function(e) {
-    if( e.key === 'Enter' ) {
-        e.preventDefault();
-        addBtn.click();
-        enableButtons();
-    }
-})
+inputArea.addEventListener('input', () => { enableButtons();})
 
 // Get items from localStorage
 
@@ -36,10 +31,13 @@ function addTask() {
     inputArea.value = inputArea.value.toLowerCase();
 
     if( !tasksArray.includes(inputArea.value) && inputArea.value ) {
+
             tasksArray.push(inputArea.value);
             renderTasks();
             inputArea.value = '';
             enableButtons();
+        
+
     } else if( tasksArray.includes(inputArea.value) ) {
         error.innerHTML = `<div class='error'>Duplicates not allowed!</div>`
     } else {
@@ -59,15 +57,45 @@ function addTask() {
 // Remove all tasks and allow restart
 
 function removeAllTasks() {
-    tasksArray = [];
-    renderTasks();
-    inputArea.value = ''
-    localStorage.clear();
-    taskArea.innerHTML = `<img src="images/giphy.webp" />
-                          <h2 class='congratulations'>Great Job!  You've finished all of your tasks!!</h2>
+
+    // Modal pop up
+
+    taskArea.innerHTML = `<div id='modal' class='modal'>Would you like to remove ALL
+                         chores?<button id='yes-btn' class='yes-btn btn-confirmation'>YES</button>
+                         <button id='no-btn' class='no-btn btn-confirmation'>NO</button></div>
                          `
+
+    document.getElementById('no-btn').addEventListener('click', function() {
+        document.getElementById('modal').classList.add('hide');
+        document.getElementById('no-btn').classList.add('hide');
+        document.getElementById('yes-btn').classList.add('hide');
+    })
+    
+    document.getElementById('yes-btn').addEventListener('click', function() {
+        tasksArray = [];
+        renderTasks();
+        inputArea.value = ''
+        localStorage.clear();
+        taskArea.innerHTML = `<img src="images/giphy.webp" />
+                            <h2 class='congratulations'>Great Job!  You've finished all of your tasks!!</h2>
+                            `
     resetBtn.classList.remove('hide');
+    })
 };   
+
+// Remove individual tasks
+
+function removeIndividualTask(index) {
+
+    tasksArray.splice(index, 1)
+
+    if( tasksArray.length === 0 ) {
+        deleteBtn.disable = true;
+        localStorage.clear();
+    }
+    console.log(deleteBtn.disable)
+    renderTasks()
+}
 
 
 // Render items to browser
@@ -82,15 +110,44 @@ function renderTasks() {
             `
     }
     taskArea.innerHTML = taskItems;
+
+    const listItems = document.querySelectorAll(".created-task");
+
+    listItems.forEach((item, index) => {
+        item.addEventListener("dblclick", function() {
+            removeIndividualTask(index);
+        })
+    })
 }
+
 
 // Conditionally disable/enable buttons
 
 function enableButtons() {
+
+    // Enable delete button if tasks are present
+
     if( tasksArray.length === 0 ) {
         deleteBtn.disabled = true;
     } else { 
         deleteBtn.disabled = false;
+    }
+
+    if( inputArea.value ) {
+        addBtn.disabled = false;
+
+        // Added 'Enter' functionality
+
+        inputArea.addEventListener('keypress', function(e) {
+
+        if( e.key === 'Enter' ) {
+            e.preventDefault();
+            addBtn.click();
+            enableButtons();
+        }
+        })
+    } else {
+        addBtn.disabled = true;
     }
 }
